trim send queue on write, decay cwnd, safe rtt
if bandwidth is constrained, low priority flows may never come up for
transmission. that could cause their send queues to grow without bound even
if waiting data expires. now at least on each write, trim the send queue of
expired messages to keep from growing without bound. unexpired data as of the
last write could still stay there forever if the flow never comes up for
transmission.

the congestion window could grow to be much larger than what's currently being
used. when this happens, decrease it a little bit on each ack so that it's a
more accurate representation of the path capacity.

before the first RTT measurement, a session's SRTT variable is supposed to
have "no value", currently stored as -1 (but should probably be NaN). either
-1 or NaN could have undesired effects if used e.g. with timers or message
lifetimes.  proved a `getSafeSRTT()` method to return a "safe" default RTT
of 200ms for this case.